Move AudioProcessingConfig into separate module
